New User Role Management
Code to add/remove the "New Green" role from new greenies with commands. I'm still setting up all the libraries and stuff on my new laptop and haven't been able to test this code yet so it almost certainly doesn't actually work yet lol.

Also I'm having trouble with git on my laptop too lol so I'm uploading this manually in the browser to check if it's a problem with my account or just a problem with my device.